split out splat-list
